# packages/shared-contracts/openapi/platform-v0.yaml
openapi: 3.1.0
info:
  title: Ekumetrics Platform API
  version: 0.0.1
  description: Contrato v0. Los eventos del agente usan /v1/ekms/events, no /events.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Liveness
      operationId: getHealth
      responses:
        "200":
          description: Servicio vivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /health/ready:
    get:
      summary: Readiness (PostgreSQL)
      operationId: getReady
      responses:
        "200":
          description: Base de datos alcanzable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Base de datos no disponible
  /v1/ekms/events:
    post:
      summary: Ingesta de envelope del agente (Fase 1)
      operationId: ingestEkmsEvents
      description: >
        Array JSON de envelope.Event. Content-Encoding gzip opcional.
        El agente deriva esta URL desde export.otlp.endpoint (puerto 4318).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/EkmsEvent"
      responses:
        "202":
          description: Lote aceptado (pendiente de implementar)
        "400":
          description: Envelope inválido
  /v1/ai/status:
    get:
      summary: Servicios y modelos de IA configurados
      operationId: getAiStatus
      responses:
        "200":
          description: Estado de EkuAssistant AI, servicios y presets de modelo
  /v1/ai/settings:
    get:
      summary: Configuracion activa de IA (sin devolver la clave)
      operationId: getAiSettings
      responses:
        "200":
          description: Servicio, modelo y si hay clave guardada
    put:
      summary: Guardar servicio, modelo y credenciales en la base
      operationId: saveAiSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service, model]
              properties:
                service:
                  type: string
                model:
                  type: string
                apiKey:
                  type: string
                baseUrl:
                  type: string
      responses:
        "200":
          description: Configuracion persistida
        "400":
          description: Servicio, modelo o credenciales invalidos
  /v1/ai/snapshot:
    get:
      summary: Snapshot de metricas del agente (Prometheus)
      operationId: getAiSnapshot
      parameters:
        - in: query
          name: agent_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hechos medidos del host
  /v1/ai/ask:
    post:
      summary: Pregunta a EkuAssistant AI
      operationId: askHolmes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                service:
                  type: string
                  description: ollama, openai, anthropic, kimi, grok, deepseek u openai_compat
                provider:
                  type: string
                  description: Alias de service (compatibilidad)
                model:
                  type: string
                  description: Id del modelo. Si se omite, se usa el default del servicio.
      responses:
        "201":
          description: Respuesta de EkuAssistant AI
        "400":
          description: Pregunta o proveedor inválido
        "503":
          description: EkuAssistant AI o el modelo no disponibles
components:
  schemas:
    HealthStatus:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
        service:
          type: string
    EkmsEvent:
      type: object
      required: [timestamp, tenant_id, site_id, agent_id, signal, source]
      properties:
        timestamp:
          type: string
          format: date-time
        tenant_id:
          type: string
        site_id:
          type: string
        agent_id:
          type: string
        asset_id:
          type: string
        asset_type:
          type: string
        vendor:
          type: string
        tier:
          type: integer
        signal:
          type: string
          enum:
            - asset_discovered
            - asset_updated
            - asset_stale
            - mac_changed
            - flow.bytes
            - flow.new
            - snmp.trap
        value:
          type: number
        unit:
          type: string
        severity:
          type: string
        source:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string